+++ src/mutations/matching/useDeleteMatching.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import API from '../api/api';
+import API from '../../api/api';
 import { useNavigate } from 'react-router-dom';
 
 const useDeleteMatching = () => {
